requirements.txt
Django==4.2.11
Requests==2.31.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
black==24.3.0
PyMySQL==1.1.1
drf-yasg==1.21.7
Faker==24.2.0
Markdown==3.3.6

